+++ src/components/ui/Header.tsx
@@ -1,56 +1,25 @@
 import { LoginArea } from "@/components/auth/LoginArea";
 import { BalanceDisplay } from "@/components/BalanceDisplay";
-import type React from "react";
-import { Link, useNavigate } from "react-router-dom";
-import { Button } from "@/components/ui/button";
+import { ClaimOnboardingTokenButton } from "@/components/ClaimOnboardingTokenButton";
 import { useCashuStore } from "@/stores/cashuStore";
 import { useOnboardingStore } from "@/stores/onboardingStore";
-import { getTokenAmount } from "@/lib/cashu";
-import { useBitcoinPrice, satsToUSD, formatUSD } from "@/hooks/useBitcoinPrice";
-import { useCurrencyDisplayStore } from "@/stores/currencyDisplayStore";
 import { useCurrentUser } from "@/hooks/useCurrentUser";
-import { DollarSign } from "lucide-react";
-import { Bitcoin } from "lucide-react";
+import type React from "react";
+import { Link } from "react-router-dom";
 
 interface HeaderProps {
   className?: string;
 }
 
 const Header: React.FC<HeaderProps> = ({ className }) => {
-  const navigate = useNavigate();
   const cashuStore = useCashuStore();
   const onboardingStore = useOnboardingStore();
-  const { showSats } = useCurrencyDisplayStore();
-  const { data: btcPrice } = useBitcoinPrice();
   const { user } = useCurrentUser();
 
-  // Check for pending onboarding token
+  // Check if we should show the claim button
   const pendingToken = cashuStore.getPendingOnboardingToken();
   const hasClaimedToken = onboardingStore.isTokenClaimed();
   const showClaimButton = user && pendingToken && !hasClaimedToken;
-
-  // Calculate the amount to display on the button
-  let claimButtonText = "Claim eCash";
-  if (showClaimButton && pendingToken) {
-    try {
-      const amount = getTokenAmount(pendingToken);
-      if (showSats) {
-        claimButtonText = `${amount.toLocaleString()} sats`;
-      } else if (btcPrice?.USD) {
-        const usd = satsToUSD(amount, btcPrice.USD);
-        if (usd !== null) {
-          claimButtonText = `${formatUSD(usd)}`;
-        }
-      }
-      claimButtonText = `Free ${claimButtonText}`;
-    } catch (error) {
-      console.error("Error calculating token amount:", error);
-    }
-  }
-
-  const handleClaimClick = () => {
-    navigate("/wallet");
-  };
 
   return (
     <div className={`flex justify-between items-center ${className || ""}`}>
@@ -61,22 +30,7 @@ const Header: React.FC<HeaderProps> = ({ className }) => {
         </h1>
       </Link>
       <div className="flex items-center gap-2">
-        {!showClaimButton && <BalanceDisplay />}
-        {showClaimButton && (
-          <Button
-            onClick={handleClaimClick}
-            size="sm"
-            variant="ghost"
-            className="animate-pulse font-medium hover:bg-green-50 dark:hover:bg-green-950"
-          >
-            {showSats ? (
-              <Bitcoin className="w-3.5 h-3.5 text-orange-500" />
-            ) : (
-              <DollarSign className="w-3.5 h-3.5 text-green-600" />
-            )}
-            {claimButtonText}
-          </Button>
-        )}
+        {showClaimButton ? <ClaimOnboardingTokenButton /> : <BalanceDisplay />}
         <LoginArea />
       </div>
     </div>
